Create my-layout.scss
Again, this is the implementation of a technology(or library) which is part of the course requirements. This is not my code. I will go back and arrange things, and I will learn from the experience. I did not code this out of thin air. This is clearly somebody else's work.